move codegen-sbt to new package
