+++ codegen-sbt/src/main/scala/caliban/codegen/sbt/CalibanPlugin.scala
@@ -1,4 +1,4 @@
-package caliban.codegen
+package caliban.codegen.sbt
 
 import sbt.Keys._
 import sbt._
