+++ codegen-sbt/src/main/scala/caliban/codegen/sbt/CalibanSourceGenerator.scala
@@ -1,4 +1,4 @@
-package caliban.codegen
+package caliban.codegen.sbt
 
 import _root_.caliban.tools._
 import sbt._
