+++ codegen-sbt/src/test/scala/caliban/codegen/OptionsParserSpec.scala
@@ -1,4 +1,4 @@
-package caliban.codegen
+package caliban.codegen.sbt
 
 import sttp.model.Header
 import zio.test.*
